[WFS provider] Extract reusable network request mechanism
Funded by Planet
