+++ src/providers/wfs/qgswfsdatasourceuri.h
@@ -16,56 +16,13 @@
 #ifndef QGSWFSDATASOURCEURI_H
 #define QGSWFSDATASOURCEURI_H
 
-#include "qgsauthmanager.h"
 #include "qgsdatasourceuri.h"
 #include "qgsrectangle.h"
 #include "qgsapplication.h"
+#include "qgsauthorizationsettings.h"
 
 #include <QNetworkRequest>
 #include <QString>
-
-// TODO: merge with QgsWmsAuthorization?
-struct QgsWFSAuthorization
-{
-  QgsWFSAuthorization( const QString &userName = QString(), const QString &password = QString(), const QString &authcfg = QString() )
-    : mUserName( userName )
-    , mPassword( password )
-    , mAuthCfg( authcfg )
-  {}
-
-  //! update authorization for request
-  bool setAuthorization( QNetworkRequest &request ) const
-  {
-    if ( !mAuthCfg.isEmpty() ) // must be non-empty value
-    {
-      return QgsApplication::authManager()->updateNetworkRequest( request, mAuthCfg );
-    }
-    else if ( !mUserName.isNull() || !mPassword.isNull() ) // allow empty values
-    {
-      request.setRawHeader( "Authorization", "Basic " + QStringLiteral( "%1:%2" ).arg( mUserName, mPassword ).toLatin1().toBase64() );
-    }
-    return true;
-  }
-
-  //! update authorization for reply
-  bool setAuthorizationReply( QNetworkReply *reply ) const
-  {
-    if ( !mAuthCfg.isEmpty() )
-    {
-      return QgsApplication::authManager()->updateNetworkReply( reply, mAuthCfg );
-    }
-    return true;
-  }
-
-  //! Username for basic http authentication
-  QString mUserName;
-
-  //! Password for basic http authentication
-  QString mPassword;
-
-  //! Authentication configuration ID
-  QString mAuthCfg;
-};
 
 /**
  * Utility class that wraps a QgsDataSourceUri with conveniency
@@ -157,7 +114,7 @@ class QgsWFSDataSourceURI
     bool hideDownloadProgressDialog() const;
 
     //! Returns authorization parameters
-    QgsWFSAuthorization &auth() { return mAuth; }
+    const QgsAuthorizationSettings &auth() const { return mAuth; }
 
     //! Builds a derived uri from a base uri
     static QString build( const QString &uri,
@@ -174,7 +131,7 @@ class QgsWFSDataSourceURI
 
   private:
     QgsDataSourceUri    mURI;
-    QgsWFSAuthorization mAuth;
+    QgsAuthorizationSettings mAuth;
     QgsStringMap mGetEndpoints;
     QgsStringMap mPostEndpoints;
 };
