+++ src/providers/wfs/qgswfssourceselect.cpp
@@ -32,6 +32,7 @@
 #include "qgssqlstatement.h"
 #include "qgssettings.h"
 #include "qgsgui.h"
+#include "qgswfsguiutils.h"
 
 #include <QDomDocument>
 #include <QListWidgetItem>
@@ -218,32 +219,10 @@ void QgsWFSSourceSelect::capabilitiesReplyFinished()
   if ( !mCapabilities )
     return;
 
-  QgsWfsCapabilities::ErrorCode err = mCapabilities->errorCode();
-  if ( err != QgsWfsCapabilities::NoError )
+  auto err = mCapabilities->errorCode();
+  if ( err != QgsBaseNetworkRequest::NoError )
   {
-    QString title;
-    switch ( err )
-    {
-      case QgsWfsCapabilities::NetworkError:
-        title = tr( "Network Error" );
-        break;
-      case QgsWfsCapabilities::XmlError:
-        title = tr( "Capabilities document is not valid" );
-        break;
-      case QgsWfsCapabilities::ServerExceptionError:
-        title = tr( "Server Exception" );
-        break;
-      default:
-        title = tr( "Error" );
-        break;
-    }
-    // handle errors
-    QMessageBox *box = new QMessageBox( QMessageBox::Critical, title, mCapabilities->errorMessage(), QMessageBox::Ok, this );
-    box->setAttribute( Qt::WA_DeleteOnClose );
-    box->setModal( true );
-    box->setObjectName( QStringLiteral( "WFSCapabilitiesErrorBox" ) );
-    if ( !property( "hideDialogs" ).toBool() )
-      box->open();
+    QgsWfsGuiUtils::displayErrorMessageOnFailedCapabilities( mCapabilities, this );
 
     emit enableButtons( false );
     return;
